requirements.txt
selenium
pytest
pytest-html
python-dotenv
pyyaml
webdriver-manager